Version 2.8. Added (uint64_t) typecast on message byte loads for better cross-compiler compatibility.

+++ prvhash42.h
@@ -31,7 +31,7 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
  * DEALINGS IN THE SOFTWARE.
  *
- * @version 2.7
+ * @version 2.8
  */
 
 //$ nocpp
@@ -97,8 +97,9 @@ inline void prvhash42( const uint8_t* const Message, const int MessageLen,
 
 	for( k = 0; k < c; k += 2 )
 	{
-		const uint64_t msg = ( k < MessageLen ? Message[ k ] : 0x100 ) |
-			( k < MessageLen - 1 ? Message[ k + 1 ] << 8 : 0x10000 );
+		const uint64_t msg =
+			( k < MessageLen ? (uint64_t) Message[ k ] : 0x100 ) |
+			( k < MessageLen - 1 ? (uint64_t) Message[ k + 1 ] << 8 : 0x10000 );
 
 		Seed *= lcg;
 		uint32_t* const hc = (uint32_t*) &Hash[ hpos ];
